Test rewriting a series of nested binds as a chain works as expected...
i.e. that Bind behaves as a monadic bind operator as intended.

It only works this way when passing a block that returns a Deferrable:
misusing #bind! like #map breaks this law and can cause confusion.
(Specifically, if a block returns something other than a Deferrable,
calling #bind! on that thing breaks us out of the monad - e.g. with a
NoMethodError - so the nested syntax doesn't really make sense.)